v0.0.9版本: 1. 修复 bug; 2. 增加批量导出导入功能; 3. 添加模版大小限制;

+++ agtms-parent/agtms-web/src/main/java/net/saisimon/agtms/web/controller/main/TemplateMainController.java
@@ -126,15 +126,9 @@ public class TemplateMainController extends MainController {
 		}
 		Long userId = AuthUtils.getUserInfo().getUserId();
 		TemplateService templateService = TemplateServiceFactory.get();
-		List<Template> templates = new ArrayList<>();
 		for (Long id : ids) {
 			Template template = templateService.getTemplate(id, userId);
-			if (templates != null) {
-				templates.add(template);
-			}
-		}
-		for (Template template : templates) {
-			if (templateService.removeTemplate(template)) {
+			if (template != null && templateService.removeTemplate(template)) {
 				templateService.dropTable(template);
 			}
 		}
